change: re-organizze report details folder
