+++ src/web/pages/reports/details/__tests__/AgentScanningHostsTab.test.tsx
@@ -8,7 +8,7 @@ import {rendererWith, screen, within} from 'web/testing';
 import Filter from 'gmp/models/filter';
 import {SEVERITY_RATING_CVSS_3} from 'gmp/utils/severity';
 import {getMockReport} from 'web/pages/reports/__fixtures__/MockReport';
-import AgentScanningHostsTab from 'web/pages/reports/details/AgentScanningHostsTab';
+import AgentScanningHostsTab from 'web/pages/reports/details/host/AgentScanningHostsTab';
 
 const filter = Filter.fromString(
   'apply_overrides=0 levels=hml rows=2 min_qod=70 first=1 sort-reverse=severity',
