+++ src/web/pages/reports/details/__tests__/CvesTab.test.tsx
@@ -9,7 +9,7 @@ import CollectionCounts from 'gmp/collection/collection-counts';
 import Filter from 'gmp/models/filter';
 import {createSession} from 'gmp/testing';
 import {getMockReport} from 'web/pages/reports/__fixtures__/MockReport';
-import CvesTab from 'web/pages/reports/details/CvesTab';
+import CvesTab from 'web/pages/reports/details/cve/CvesTab';
 
 const filter = Filter.fromString(
   'apply_overrides=0 levels=hml rows=2 min_qod=70 first=1 sort-reverse=severity',
