+++ src/web/pages/reports/details/__tests__/CvesTable.test.tsx
@@ -9,7 +9,7 @@ import CollectionCounts from 'gmp/collection/collection-counts';
 import Filter from 'gmp/models/filter';
 import type {ReportActiveCve} from 'gmp/models/report/parser';
 import {SEVERITY_RATING_CVSS_3} from 'gmp/utils/severity';
-import CvesTable from 'web/pages/reports/details/CvesTable';
+import CvesTable from 'web/pages/reports/details/cve/CvesTable';
 
 const filter = Filter.fromString('first=1 rows=10');
