+++ src/web/pages/reports/details/application/ApplicationsTable.tsx
