+++ src/web/pages/reports/details/cve/ClosedCvesTable.tsx
