+++ src/web/pages/reports/details/cve/CvesTab.tsx
@@ -14,9 +14,9 @@ import {
   NO_RELOAD,
   USE_DEFAULT_RELOAD_INTERVAL_ACTIVE,
 } from 'web/components/loading/Reload';
-import useGetReportClosedCves from 'web/hooks/use-query/report-closed-cves';
+import useGetReportCves from 'web/hooks/use-query/report-cves';
 import useFilterSortBy from 'web/hooks/useFilterSortBy';
-import ClosedCvesTable from 'web/pages/reports/details/ClosedCvesTable';
+import CvesTable from 'web/pages/reports/details/cve/CvesTable';
 import ReportEntitiesContainer from 'web/pages/reports/details/ReportEntitiesContainer';
 import {
   makeCompareIp,
@@ -24,46 +24,42 @@ import {
   makeCompareSeverity,
 } from 'web/utils/Sort';
 
-interface ClosedCvesTabProps {
+interface CvesTabProps {
   filter?: Filter;
   reportId: string;
   status: TaskStatus;
 }
 
-export const closedCvesSortFunctions = {
+export const cvesSortFunctions = {
   cve: makeCompareString('cveId'),
   host: makeCompareIp(entity => entity.host.ip),
   nvt: makeCompareString(entity => entity.source?.description),
   severity: makeCompareSeverity(),
 };
 
-const ClosedCvesTabWrapper = ({
-  filter,
-  reportId,
-  status,
-}: ClosedCvesTabProps) => {
+const CvesTabWrapper = ({filter, reportId, status}: CvesTabProps) => {
   const [_] = useTranslation();
 
   const baseFilter = useMemo(() => {
     return isDefined(filter) ? filter.copy() : new Filter();
   }, [filter]);
 
-  const [closedCvesFilter, setClosedCvesFilter] = useState<Filter>(baseFilter);
+  const [cvesFilter, setCvesFilter] = useState<Filter>(baseFilter);
 
-  const {data, isLoading, isFetching, isError, error} = useGetReportClosedCves({
+  const {data, isLoading, isFetching, isError, error} = useGetReportCves({
     reportId,
-    filter: closedCvesFilter,
+    filter: cvesFilter,
     refetchInterval: isActive(status)
       ? USE_DEFAULT_RELOAD_INTERVAL_ACTIVE
       : NO_RELOAD,
   });
 
   const updateFilter = (newFilter: Filter) => {
-    setClosedCvesFilter(newFilter);
+    setCvesFilter(newFilter);
   };
 
   const [sortBy, sortDir, handleSortChange] = useFilterSortBy(
-    closedCvesFilter,
+    cvesFilter,
     updateFilter,
   );
 
@@ -71,17 +67,16 @@ const ClosedCvesTabWrapper = ({
     return (
       <ErrorPanel
         error={error}
-        message={_('Error while loading Closed CVEs for Report {{reportId}}', {
+        message={_('Error while loading CVEs for Report {{reportId}}', {
           reportId,
         })}
       />
     );
   }
 
-  const {entities: closedCves = [], entitiesCounts: closedCvesCounts} =
-    data || {};
+  const {entities: cves = [], entitiesCounts: cvesCounts} = data || {};
 
-  const displayedFilter = closedCvesFilter;
+  const displayedFilter = cvesFilter;
 
   if (isLoading && !data) {
     return <Loading />;
@@ -89,11 +84,11 @@ const ClosedCvesTabWrapper = ({
 
   return (
     <ReportEntitiesContainer
-      counts={closedCvesCounts}
-      entities={closedCves}
+      counts={cvesCounts}
+      entities={cves}
       filter={displayedFilter}
       sortField={sortBy || 'severity'}
-      sortFunctions={closedCvesSortFunctions}
+      sortFunctions={cvesSortFunctions}
       sortReverse={sortDir === 'asc'}
     >
       {({
@@ -106,8 +101,8 @@ const ClosedCvesTabWrapper = ({
         onNextClick,
         onPreviousClick,
       }) => (
-        <ClosedCvesTable
-          // @ts-expect-error entities are ReportClosedCve[], not Model[]
+        <CvesTable
+          // @ts-expect-error entities are ReportActiveCve[], not Model[]
           entities={entities}
           entitiesCounts={entitiesCounts}
           filter={displayedFilter}
@@ -126,4 +121,4 @@ const ClosedCvesTabWrapper = ({
   );
 };
 
-export default ClosedCvesTabWrapper;
+export default CvesTabWrapper;
