+++ src/web/pages/reports/details/cve/CvesTable.tsx
