+++ src/web/pages/reports/details/error/ErrorsTable.tsx
