+++ src/web/pages/reports/details/host/AgentScanningHostsTab.tsx
@@ -6,7 +6,7 @@
 import type CollectionCounts from 'gmp/collection/collection-counts';
 import type Filter from 'gmp/models/filter';
 import type ReportHost from 'gmp/models/report/host';
-import AgentScanningHostsTable from 'web/pages/reports/details/AgentScanningHostsTable';
+import AgentScanningHostsTable from 'web/pages/reports/details/host/AgentScanningHostsTable';
 import ReportEntitiesContainer from 'web/pages/reports/details/ReportEntitiesContainer';
 import {
   makeCompareIp,
