+++ src/web/pages/reports/details/host/AgentScanningHostsTable.tsx
