+++ src/web/pages/reports/details/host/HostsTable.tsx
