+++ src/web/pages/reports/details/operating-system/OperatingSystemsTable.tsx
