+++ src/web/pages/reports/details/result/ContainerScanningResultsTable.tsx
